Restore Environment.GetEnvironmentVariable() on UWP
Sets things up for System.Runtime.Extensions.Tests bringup
on ILC.

Creates platform-specific EnvironmentAugments files.

Move GetEnvironmentVariable() over to EnvironmentAugments
(not used within CoreLib.)

Get of rid of emasculated UWP version and make it
use regular Windows version.
